AdoraChat/requirements.txt
Flask
python-dotenv
flask-firebase